fixed a few compiler warnings in visual studio.
git-svn-id: https://libgdf.svn.sourceforge.net/svnroot/libgdf/trunk@27 00a3b082-7228-46c3-a6fa-0e10d54cf57f

+++ libgdf/src/GDFHeaderAccess.cpp
@@ -435,6 +435,6 @@ namespace gdf
         if( it == m_sighdr.end() )
             return 0;
 
-        return m_mainhdr.get_num_datarecords() * it->second.get_samples_per_record();
+		return boost::numeric_cast<size_t>( m_mainhdr.get_num_datarecords() ) * it->second.get_samples_per_record();
     }
 }
